feat: add visual confirmation that you copied link
